pyproject.toml
[project]
name = "seekit"
version = "0.0.2"
authors = [{ name = "lexiforest", email = "infinitesheldon@gmail.com" }]
description = "Python API for various search engines"
license = "MIT"
dependencies = [
    "curl-cffi>=0.14.0",
    "fastapi>=0.110.0,<1.0",
    "lxml_html_clean>=0.4.2",
    "lxml>=5.3.2",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "rich>=13.9.0",
    "uvicorn>=0.29.0,<1.0",
]
readme = "README.md"
requires-python = ">=3.10"
urls = { "repository" = "https://github.com/riverside-ai/seekit" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

[project.scripts]
seekit = "seekit.cli:main"
seekit-server = "seekit.server:main"


[project.optional-dependencies]
extra = [
    "readability-lxml>=0.8.1",
    "markdownify>=1.1.0",
]
dev = [
    "charset_normalizer>=3.3.2,<4.0",
    "coverage>=6.4.1,<7.0",
    "cryptography>=42.0.5,<43.0",
    "httpx==0.23.1",  # don't change, tests will raise "httpx.InvalidURL: Invalid URL component 'path'"
    "mypy>=1.9.0,<2.0",
    "pytest>=8.1.1,<9.0",
    "pytest-asyncio>=0.23.6,<1.0",
    "pytest-trio>=0.8.0,<1.0",
    "ruff>=0.3.5,<1.0",
    "trio>=0.25.0,<1.0",
    "trustme>=1.1.0,<2.0",
    "uvicorn>=0.29.0,<1.0",
    "websockets>=12.0,<13.0",
    "typing_extensions",
]
build = [
    "cibuildwheel",
    "wheel",
]
test = [
    "charset_normalizer>=3.3.2,<4.0",
    "cryptography>=42.0.5,<43.0",
    "fastapi==0.110.0,<1.0",
    "httpx==0.23.1",  # don't change, tests will raise "httpx.InvalidURL: Invalid URL component 'path'"
    "proxy.py>=2.4.3,<3.0",
    "pytest>=8.1.1,<9.0",
    "pytest-asyncio>=0.23.6,<1.0",
    "pytest-trio>=0.8.0,<1.0",
    "python-multipart>=0.0.9,<1.0",
    "trio>=0.25.0,<1.0",
    "trustme>=1.1.0,<2.0",
    "uvicorn>=0.29.0,<1.0",
    "websockets>=12.0,<13.0",
    "typing_extensions",
]
docs = [
    "sphinx>=8.0.0,<9.0",
]


[build-system]
requires = ["wheel", "setuptools", "cffi>=1.12.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]


[tool.pytest.ini_options]
# pythonpath = [ "." ]
asyncio_mode = "auto"


[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # Pyflakes
    "UP",  # pyupgrade
    "B",  # flake8-bugbear
    "SIM",  # flake8-simplify
]
ignore = [
    "UP007"
]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
#warn_unused_ignores = true
#strict = true
exclude = ["docs/", "examples/", "scripts/", "tests/", "build/"]
